machine_server/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
sqlalchemy>=2.0.23
requests>=2.31.0
google-genai>=0.1.2
python-dotenv>=1.0.0
pytz>=2023.3.post1
pandas>=2.1.3
pytz
pandas
google-genai
